+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @convergence-rfq/cli
 
+## 4.0.17
+
+### Patch Changes
+
+- add method createEuropeanProgram
+- Updated dependencies
+  - @convergence-rfq/sdk@4.0.17
+
+## 4.0.16
+
+### Patch Changes
+
+- export initializeNewOptionMeta fn
+- Updated dependencies
+  - @convergence-rfq/sdk@4.0.16
+
+## 4.0.15
+
+### Patch Changes
+
+- add initializeNewOptionMeta method
+- Updated dependencies
+  - @convergence-rfq/sdk@4.0.15
+
 ## 4.0.14
 
 ### Patch Changes