+++ packages/js/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @convergence-rfq/sdk
 
+## 4.0.17
+
+### Patch Changes
+
+- add method createEuropeanProgram
+
+## 4.0.16
+
+### Patch Changes
+
+- export initializeNewOptionMeta fn
+
+## 4.0.15
+
+### Patch Changes
+
+- add initializeNewOptionMeta method
+
 ## 4.0.14
 
 ### Patch Changes